% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday_newchanges.R
\name{run_birthday_paradox_app}
\alias{run_birthday_paradox_app}
\title{Probability of Repeated Outcomes in Random Events Shiny App}
\usage{
run_birthday_paradox_app()
}
\arguments{
\item{none}{No parameters required}
}
\value{
A Shiny app interface to explore probabilities and simulate events
}
\description{
This Shiny app allows users to explore the probabilities of repeated outcomes
in various random events such as the Birthday Paradox, Dice Rolls, Card Draws, and Coin Flips.
It also includes a simulation feature to visualize the outcomes and highlight repeated events.
}
\examples{
run_probability_app()  # Launch the app

}
