% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_lln_visualizer.R
\name{run_clt_lln_app}
\alias{run_clt_lln_app}
\title{Statistical Concepts Explorer: CLT and LLN}
\usage{
run_clt_lln_app()
}
\description{
This Shiny app allows users to explore two important statistical concepts:
the Central Limit Theorem (CLT) and the Law of Large Numbers (LLN). Users
can select either concept, choose parameters for different distributions or
experiments, and visualize the respective results.
}
\examples{
run_clt_lln_app() # This function runs the Shiny app.

}
