% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_lln_visualizer.R
\name{run_clt_wlln_app}
\alias{run_clt_wlln_app}
\title{Statistical Concepts Explorer: CLT and WLLN}
\usage{
run_clt_wlln_app()
}
\description{
This Shiny app allows users to explore two foundational statistical concepts:
the Central Limit Theorem (CLT) and the Weak Law of Large Numbers (WLLN).
Users can select either concept, choose parameters for different distributions
or experiments, and visualize the respective results.
}
\examples{
\dontrun{
run_clt_wlln_app() # This function runs the Shiny app.
}

}
