% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_test.R
\name{run_continuous_app}
\alias{run_continuous_app}
\title{Run Continuous Sample Space Probability Calculator App}
\usage{
run_continuous_app()
}
\value{
The Shiny app is launched in the default web browser.
}
\description{
This function launches the Shiny app for calculating probabilities in continuous sample spaces.
Users can select between Uniform and Normal distributions, define variable ranges and intervals,
and calculate both joint and individual probabilities. The app also provides visualizations
of the defined ranges and intervals.
}
\examples{
if (interactive()) {
  run_continuous_app()
}
}
