% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{run_descriptive_analysis_app}
\alias{run_descriptive_analysis_app}
\title{Run Descriptive Analysis Shiny App}
\usage{
run_descriptive_analysis_app()
}
\value{
The Shiny app is launched in the default web browser.
}
\description{
This function launches the Shiny app for descriptive analysis of datasets.
The app allows users to upload Excel files, select datasets (including built-in datasets like `iris` and `palmerpenguins`),
and create various plots such as histograms, boxplots, and scatter plots.
}
\examples{
if (interactive()) {
  run_descriptive_analysis_app()
}
}
