% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_test.R
\name{run_discrete_app}
\alias{run_discrete_app}
\title{Launch the Probability Calculator with Operations Shiny App}
\usage{
run_discrete_app()
}
\value{
A Shiny web application object that runs locally in the user's web browser.
}
\description{
The \code{run_discrete_app} function initializes and launches a Shiny web application
designed to calculate and visualize probabilities based on discrete outcomes and
events. The app allows users to define outcomes, events, and trials, input
probabilities for each outcome, and perform operations such as union,
intersection, and complement on the defined events. The result of the operations,
along with the event probabilities, is then displayed for the user.
}
\details{
The app provides an interactive interface where users can:
\itemize{
\item Define the number of outcomes, events, and trials.
\item Input the probability for each outcome.
\item Assign outcomes to each trial within an event.
\item Perform operations on the events (union, intersection, complement).
\item View the calculated probabilities for each event and the result of the selected operation.
}
}
\examples{
if (interactive()) {
  run_discrete_app()
}

}
