% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{run_distribution_visualizer_app}
\alias{run_distribution_visualizer_app}
\title{Run Distributions Visualizer App}
\usage{
run_distribution_visualizer_app()
}
\value{
The Shiny app is launched in the default web browser.
}
\description{
This function launches the Shiny app for visualizing many distributions with continuous and discrete sample spaces.
Users can select between Uniform, Normal, Exponential and Binomial distributions, define variable ranges and intervals,
and calculate both joint and individual probabilities. The app also provides visualizations
of the defined ranges and intervals.
}
\examples{
if (interactive()) {
  run_distribution_visualizer_app()
}
}
