% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_probabilty_tree.R
\name{run_probability_tree_app}
\alias{run_probability_tree_app}
\title{Run the Probability Tree Shiny App}
\usage{
run_probability_tree_app()
}
\description{
This function launches a Shiny application that visualizes a probability tree
for disease prevalence, test sensitivity, and test specificity.
}
\examples{
\dontrun{
run_probability_tree_app() # Launch the app
}
}
