% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_experiments.R
\name{run_three_experiments_app}
\alias{run_three_experiments_app}
\title{Run the Three Events Visualization Shiny App}
\usage{
run_three_experiments_app()
}
\value{
A Shiny app instance that runs locally in your browser.
}
\description{
This function launches a Shiny app that visualizes the outcomes of three experiments (Coin Toss or Dice Roll).
Users can select experiment types and outcomes, and visualize probabilities and selected outcomes in a 3D plot.
}
\details{
Features:
- Select between "Coin Toss" or "Dice Roll" for three different experiments.
- Choose specific outcomes for each event.
- Visualize the probability of selected outcomes and display a 3D scatter plot of possible results.
}
\examples{
if (interactive()) {
  run_three_experiments_app()
}

}
