% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_trials_final.R
\name{runThreeEventsApp}
\alias{runThreeEventsApp}
\title{Run the Three Events Visualization Shiny App}
\usage{
runThreeEventsApp()
}
\value{
A Shiny app instance that runs locally in your browser.
}
\description{
This function launches a Shiny app that visualizes the outcomes of three events (Coin Toss or Dice Roll).
Users can select event types and outcomes, and visualize probabilities and selected outcomes in a 3D plot.
}
\details{
Features:
\itemize{
\item Select between "Coin Toss" or "Dice Roll" for three different events.
\item Choose specific outcomes for each event.
\item Visualize the probability of selected outcomes and display a 3D scatter plot of possible results.
}
}
\examples{
if (interactive()) {
  runThreeEventsApp()
}

}
